Move letter remote access to sbc_compassion
Move fields related to remote access of letters to
SponsorshipCorrespondence in the sbc_compassion module. This way, the same
fields can be used for accessing all types of letters remotely.
